fix: fix graph for settlement contract trades (#492)
* fix: do not overwrite contract alias if it's also a trader

* refactor: remove default export

* fix: error handling when graph is failed to load

* refactor: extracted TransactionBatchGraph/index logic onto hooks and utils
